pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docmirror"
version = "0.1.0"
description = "Universal document parsing engine with OCR, layout analysis, and table extraction"
requires-python = ">=3.10"
license = "Apache-2.0"
readme = "README.md"
keywords = [
    "document-parsing",
    "ocr",
    "pdf",
    "table-extraction",
    "layout-analysis",
    "document-ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
]

[project.urls]
Homepage = "https://github.com/docmirror/docmirror"
Documentation = "https://docmirror.readthedocs.io"
Repository = "https://github.com/docmirror/docmirror"
Issues = "https://github.com/docmirror/docmirror/issues"
Changelog = "https://github.com/docmirror/docmirror/blob/main/CHANGELOG.md"

dependencies = [
    "pydantic>=2.0",
    "filetype>=1.2",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
pdf = ["PyMuPDF>=1.23", "pdfplumber>=0.10"]
ocr = ["rapidocr-onnxruntime>=1.3", "opencv-python-headless>=4.8", "numpy>=1.24"]
layout = ["rapid-layout>=0.1"]
table = ["rapid-table>=0.2"]
formula = ["rapid-latex-ocr>=0.0.6"]
office = ["python-docx>=1.0", "openpyxl>=3.1", "python-pptx>=0.6"]
vlm = ["httpx>=0.25"]
security = ["pikepdf>=8.0"]
cache = ["redis>=5.0"]
langdetect = ["fast-langdetect>=0.2"]
all = [
    "docmirror[pdf,ocr,layout,table,formula,office,vlm,security,cache,langdetect]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
    "coverage>=7.4",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
